Refactor av fagsaker til egen mappe.
