+++ src/reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'react-redux-form-validation';
 
-import fagsakerReducer from './ducks/fagsaker';
+import fagsakerReducer from './ducks/fagsaker/fagsaker';
 import soknadReducer from './ducks/soknad';
 import faktaavklaringReducer from './ducks/faktaavklaring';
 import vurderingReducer from './ducks/vurdering';
